refactor: remove the types module
