+++ src/chip/family/mem_region.rs
@@ -1,5 +1,5 @@
 #[derive(Clone, Debug)]
-pub(crate) struct MemRegion {
+pub struct MemRegion {
     pub flash_origin: usize,
     pub flash_length: usize,
 
@@ -8,38 +8,38 @@ pub(crate) struct MemRegion {
 }
 
 impl MemRegion {
-    pub(crate) const NRF52805: Self = Self {
+    pub const NRF52805: Self = Self {
         flash_origin: 0,
         flash_length: 192,
         ram_origin: 0x2 << 28,
         ram_length: 24,
     };
-    pub(crate) const NRF52810: Self = Self::NRF52805;
-    pub(crate) const NRF52811: Self = Self::NRF52805;
+    pub const NRF52810: Self = Self::NRF52805;
+    pub const NRF52811: Self = Self::NRF52805;
 
-    pub(crate) const NRF52820: Self = Self {
+    pub const NRF52820: Self = Self {
         flash_origin: 0,
         flash_length: 256,
         ram_origin: 0x2 << 28,
         ram_length: 32,
     };
 
-    pub(crate) const NRF52832_XXAA: Self = Self {
+    pub const NRF52832_XXAA: Self = Self {
         flash_origin: 0,
         flash_length: 512,
         ram_origin: 0x2 << 28,
         ram_length: 64,
     };
-    pub(crate) const NRF52832_XXAB: Self = Self::NRF52820;
+    pub const NRF52832_XXAB: Self = Self::NRF52820;
 
-    pub(crate) const NRF52833: Self = Self {
+    pub const NRF52833: Self = Self {
         flash_origin: 0,
         flash_length: 512,
         ram_origin: 0x2 << 28,
         ram_length: 128,
     };
 
-    pub(crate) const NRF52840: Self = Self {
+    pub const NRF52840: Self = Self {
         flash_origin: 0,
         flash_length: 1024,
         ram_origin: 0x2 << 28,
