+++ src/parser.rs
@@ -1,5 +1,7 @@
-use crate::types::init_args::InitArgs;
+pub mod init_args;
+
 use clap::{Args, Parser, Subcommand};
+use init_args::InitArgs;
 
 #[derive(Parser)]
 #[command(name = "cargo")]
