--- src/types.rs
@@ -1,4 +1,0 @@
-pub mod chip;
-pub mod error;
-pub mod init_args;
-pub mod parser;
